pyproject.toml
[tool.poetry]
name = "vivlio"
version = "0.1.0"
description = "Business ready documents from Meraki"
authors = ["John Capobianco <ptcapo@gmail.com>"]
packages = [{ include = "vivlio", from = "src"}]
readme = "README.md"

[tool.poetry.scripts]
vivlio = "vivlio.script:run"

[tool.poetry.dependencies]
python = "^3.8"
requests = "^2.28.2"
rich-click = "^1.6.1"
pyyaml = "^6.0"
jinja2 = "^3.1.2"
aiohttp = "^3.8.4"
aiofiles = "^23.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
